requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-multipart==0.0.20
gspread==6.1.4
google-auth==2.38.0
